requirements.txt
pypdf>=4.0.0
pdfminer.six>=20221105
playwright>=1.40.0
anthropic>=0.25.0
mcp>=1.0.0
yt-dlp>=2024.1.0
openai-whisper>=20231117
Pillow>=10.0.0
pytest>=8.0.0
